requirements.txt
fastapi
uvicorn
pydantic
numpy
faiss-cpu
sentence-transformers
rank-bm25
nltk
PyPDF2
